# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  dto.ChangePasswordRequest:
    properties:
      current_password:
        type: string
      new_password:
        minLength: 6
        type: string
    required:
    - current_password
    - new_password
    type: object
...
  dto.ErrorResponse:
    properties:
      code:
        type: integer
      details:
        type: string
      message:
        type: string
    type: object
...
  dto.SuccessResponse:
    properties:
      data: {}
      message:
        type: string
    type: object
...
  dto.UserListResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/dto.UserResponse'
        type: array
      page:
        type: integer
      page_size:
        type: integer
      total_count:
        type: integer
      total_pages:
        type: integer
    type: object
  dto.UserRequest:
    properties:
      branch_id:
        type: string
      email:
        type: string
      name:
        type: string
      password:
        type: string
      role:
        type: string
    required:
    - email
    - name
    - role
    type: object
  dto.UserResponse:
    properties:
      branch_id:
        type: string
      created_at:
        type: string
      email:
        type: string
      id:
        type: string
      last_login_at:
        type: string
      name:
        type: string
      role:
        type: string
      status:
        type: string
      tenant_id:
        type: string
      updated_at:
        type: string
    type: object
...
host: localhost:8084
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: API para o sistema ERP de supermercados
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: SuperERP API
  version: "1.0"
paths:
...
  /setup/admin:
    post:
      consumes:
      - application/json
      description: Cria o primeiro usuário administrador para um tenant (não requer
        autenticação)
      parameters:
      - description: ID do tenant
        in: header
        name: tenant-id
        required: true
        type: string
      - description: Dados do usuário administrador
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/dto.UserRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.UserResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: Cria o primeiro usuário administrador
      tags:
      - setup
...
  /users:
    get:
      description: Lista os usuários do tenant com paginação
      parameters:
      - description: ID do tenant
        in: header
        name: tenant-id
        required: true
        type: string
      - description: Página
        in: query
        name: page
        type: integer
      - description: Itens por página
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.UserListResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: Lista os usuários
      tags:
      - users
    post:
      consumes:
      - application/json
      description: Cria um novo usuário no sistema
      parameters:
      - description: ID do tenant
        in: header
        name: tenant-id
        required: true
        type: string
      - description: Dados do usuário
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/dto.UserRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.UserResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: Cria um novo usuário
      tags:
      - users
  /users/{id}:
    delete:
      description: Remove um usuário do sistema
      parameters:
      - description: ID do tenant
        in: header
        name: tenant-id
        required: true
        type: string
      - description: ID do usuário
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.SuccessResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: Remove um usuário
      tags:
      - users
    get:
      description: Busca um usuário pelo seu ID
      parameters:
      - description: ID do tenant
        in: header
        name: tenant-id
        required: true
        type: string
      - description: ID do usuário
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.UserResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: Busca um usuário pelo ID
      tags:
      - users
    put:
      consumes:
      - application/json
      description: Atualiza os dados de um usuário existente
      parameters:
      - description: ID do tenant
        in: header
        name: tenant-id
        required: true
        type: string
      - description: ID do usuário
        in: path
        name: id
        required: true
        type: string
      - description: Dados do usuário
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/dto.UserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.UserResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: Atualiza um usuário
      tags:
      - users
  /users/{id}/password:
    patch:
      consumes:
      - application/json
      description: Permite que um usuário altere sua própria senha
      parameters:
      - description: ID do tenant
        in: header
        name: tenant-id
        required: true
        type: string
      - description: ID do usuário
        in: path
        name: id
        required: true
        type: string
      - description: Dados para troca de senha
        in: body
        name: password
        required: true
        schema:
          $ref: '#/definitions/dto.ChangePasswordRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: Altera a senha de um usuário
      tags:
      - users
  /users/{id}/status/{status}:
    patch:
      description: Atualiza o status de um usuário (ativo/inativo/bloqueado)
      parameters:
      - description: ID do tenant
        in: header
        name: tenant-id
        required: true
        type: string
      - description: ID do usuário
        in: path
        name: id
        required: true
        type: string
      - description: Novo status (active/inactive/blocked)
        in: path
        name: status
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.UserResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: Atualiza o status de um usuário
      tags:
      - users
  /users/branch/{branch_id}:
    get:
      description: Lista os usuários de uma filial específica com paginação
      parameters:
      - description: ID do tenant
        in: header
        name: tenant-id
        required: true
        type: string
      - description: ID da filial
        in: path
        name: branch_id
        required: true
        type: string
      - description: Página
        in: query
        name: page
        type: integer
      - description: Itens por página
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.UserListResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: Lista os usuários de uma filial
      tags:
      - users
securityDefinitions:
  Bearer:
    description: 'Cabeçalho de autenticação JWT usando o esquema Bearer. Exemplo:
      "Bearer {token}"'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"